+++ ui/app/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "2.5.7";
+export const APP_VERSION = "2.5.8";